BLAST: Handle detached layers.
detachChildren and mLayerDetached were implemented to avoid geometry updates
to layers in the cases of animation. For BufferQueue layers buffers are still
consumed as normal. This is useful as otherwise the client could get hung
inevitably waiting on a buffer to return. We recreate this semantic for BufferState
even though it is a little in-consistent. detachChildren is shortly slated for removal
by the hierarchy mirroring work so we don't need to worry about it too much.

Bug: 135786080
Test: Manual
Change-Id: I86c78fecc921218582918dda1ee7d8ee6c48f922