SF: Clear layers properly doing client composition
When performing client composition, layers that are set to use device
composition should clear the client composition framebuffer unless they
are the bottom most layer.

This regressed when I moved the code to CompositionEngine, and I caught
it while adding a unit test.

The expected behavior is now covered by a new test.

Bug: 121291683
Test: atest libcompositionengine_test
Test: YouTube, Twitch PiP on crosshatch

Change-Id: Id0db7b0d21afadc331d9b42dadef1d90a989389e