Fix touch can't work after enabled the associated viewport
If a device has associated port but there is no viewport found, it would
be expeced to be disabled. But for some devices, such as touch screen or
joystick, they have to read the axis ranges info in InputMapper at first
time configuration changed. So we have to defer disabling the device after
info has been read when the device plugged.

Bug: 140205788
Test: atest inputflinger_tests
Change-Id: I9dd55e0016b6a020aab211dada45880195aec8dd