Use static bool instead of define
The defines for debugging prevent compilers from compiling the code, as
well as make the code harder to read due to identation.
Using static bools helps ensure that the code inside the debugging logic
compiles, and a decent compiler should be able to remove the dead code
anyways.

Test: adb logcat and inspect the InputDispatcher logs for when debugging
is true
Bug: 70668286

Change-Id: If1a057da4193f965e35d2ca7c1bf39249b9222fa