Corrected drawingTransformMatrix
The transformMatrix needed to be calculated in prepareClientComposition
so storing it in BufferInfo was incorrect. Instead, allow the transformMatrix
to get calculated when prepareClientComposition but avoid using the consumer
for BufferQueueLayer.

Also combined getDrawingTransformMatrix and setFilteringEnabled since
setFilteringEnabled was only used to update the transformMatrix before getting
it. BufferQueueLayers no longer go through the consumer to get the
transformMatrix and just directly calculate the matrix using the GLUtils class.

Fixes: 140759442
Test: atest android.view.cts.ASurfaceControlTest
Change-Id: I09179de275bd905960fb366e45e36fb9874bc0b3