Crash if too many open files
If there are too many open files, and we can't dup a socket because we
received EMFILE or ENFILE, let's crash immediately to make this problem
obvious.
Otherwise, we are simply propagating null pointers through the system
and witness null pointer dereference later (in the good cases), or
undefined behaviour.

Bug: 141111452
Test: none
Change-Id: I012bc050e5246456e00c22f0c17a5752b2708875