GraphicBuffer::lock Convert default parameters to explicitely library calls.
After commit d2f4dafa520baef942ed2937f8a328018f3fc72d
Adding lock function that obtains bytesPerPixel and bytesPerStride
the GraphicBuffer::lock function is extended with new parameters
bytesPerPixel and bytesPerStride. To remain compatibility these
parameters where marked as default arguments, which initializes with
nullptr as value.
This is done on compile level and adapts the calls of this method.
That means that the resulting library contains only the method
signature which contains the two additional parameters.

Blobs which accesses this library searching it with the "old" signature,
where they compiled against. Normally it is sufficient to shim the
library and add the missing function call. This helps to resolve
the dependency on linker level but crashes when using this function.

Simple debug statements which logs the parameter values of the shimed
library overriden function but also on the GraphicBuffer::lock method
printing some weird values. So something fundamental is wrong here when
using shims.

Change-Id: I979b3d7da67c8b7c95c761fb2ef6a48bd3750886