Ensure input has CAP_BLOCK_SUSPEND
The process that's running the input code (currently, SYSTEM_SERVER)
should be able to prevent suspend from happening. This is to make sure
that we can always process the touch events.

Since the kernel is not able to enforce the permissions for
CAP_BLOCK_SUSPEND when EPOLLWAKEUP is being used, and just silently
fails (see http://man7.org/linux/man-pages/man2/epoll_ctl.2.html and
specifically the "BUGS" section), we should really make sure that the
current process is able to hold the wakelock.

Bug: 141532209
Test: boot up the phone and ensure it doesn't crash
Change-Id: Ibde183cf7a41dd5369873093be373c10dd580115