Hold wakelock while handling device add/remove
EventHub is using epoll (with inotify) to monitor for input devices
being added or removed. This is done by setting a watch on the directory
/dev/input. After the device is added or removed, epoll_wait would
return, and the code would proceed to handle the event.

To ensure that the system is awake the entire time an input device is
being added or removed, use EPOLLWAKEUP for the inotify fd.

Bug: 141532209
Test: none
Change-Id: Ie23ef459bdecdf1df929c34781165e63cae313bf