binder_parcel_fuzzer: fuzz parcle types separately
Before, binder, hwbinder, and binder_ndk parcels were always checked
against the same instuctions/data. However, because we are using guided
fuzzing, an interesting case for one parcel is highly likely to be a
waste of time for the other parcels. By only fuzzing one at a time, we
give the fuzzing engine the ability to focus in on individual parcel
implementations.

Bug: 131861045
Test: binder_parcel_fuzzer (and verify all backends are hit)
Change-Id: Ifd3e75828e68eb55d7cebfd3b40a31b0192c9991