Set isRelativeOf flag when calling setZOrderRelativeOf
The current code sets the isRelativeFlag to true when setRelativeLayer
is called and false when setLayer is called. However, those two requests
are only coming from client requests. There are places where zRelativeOf
is removed for other reasons. Therefore, set isRelativeOf flag when
calling setZOrderRelativeOf based on whether relativeOf is null or not.
This will ensure it covers all places that zRelativeOf is updated.

Fixes: 142534793
Test: RelativeZTest.LayerAndRelativeRemoved
Change-Id: I338592f2e2e88598f2d8a4c0957b551da310a4bf