Look up Connection by token
Currently, we are passing in the entire input channel in order to find
the corresponding connection. But if an input channel is uniquely
identified by the token, we should only have to pass in the token.

In some cases, we are even doing the circular lookup: find input channel
by token, then use inputchannel->token to find the corresponding
connection.

We should be able to unregisterInputChannel by token instead of by
inputchannel.

Bug: 142581626
Test: basic interaction with the device
Change-Id: Ie438c5136186f9dd14df6df08aaabeb35a336f05