Revert "Add 2 addtional VkFormat's to vkGetPhysicalDeviceSurfaceFormatsKHR"
This reverts commit 94ace213b8369237726860a20cafb9d80fd52dd8.

The Android Vulkan loader has never queried what formats are available for use
with swapchains; but has always hard-coded a list of formats.  All initial
Vulkan 1.0/1.1 drivers on Android supported those formats in order to work.
The change being reverted added to the hard-coded list, but not all drivers
support those formats for use by swapchains.  A future extension will be
created to allow the swapchain code to query which formats can be supported.

Bug: b/124942437
Test: mm