Use /data/media path when migrating legacy OBB data
The MediaProvider process is running as the FUSE daemon and is handling
IO requests from the kernel on /sdcard. Accessing a path on /sdcard
can hang forever if the daemon is not responding.

At user 0 startup (boot), the system_server waits for migration of OBB
data and it can deadlock if the FUSE daemon is not running.

To break this cyclic dependency between the FUSE daemon and the
system_server, we now migrate OBB using /data/media instead of
/sdcard, thereby bypassing the FUSE daemon.

Test: adb shell sm set-virtual-disk true
&& adb shell sm partition disk:7,112 private && adb reboot
&& adb shell ls /sdcard

Bug: 135341433
Change-Id: Ib64f7eb7608704e1b53a1b923cc068172fe05e1c