Remove HWC_Queued trace event
The latch event inherently indicates that the frame is queued for hwc
composing. It would be repetitive to show this event as well.
Additionally, the point at which this event is currently emitted is also
wrong since this can show up repeatedly if the GPU workload is too much
and the acquirefence takes time to be signaled.

Test: Take a trace with GAPID and check if HWC_Queued is emitted or not
Bug: 144806504
Change-Id: I3c2e65a74c318e05c975975da0c3b0f5c92265d7