Add toString and remove default case
For a future CL, we will need to print some bools. To facilitate that,
add a "toString" function.

Also, if we add a focus type, it would be good that the compiler
highlights cases where the new type isn't being handled in a switch
statement. Therefore, remove all default cases in switch statements
where feasible.

Because we are now adding an "unused" function that compiler complains
about, also convert DEBUG_TRANSPORT_ACTIONS into a constexpr bool.

Bug: none
Test: presubmit
Change-Id: Ie8472707f1699d2009232b9f36abae9a36a97f5f