Set the clock id in FrameTracer for TracePacket
Perfetto uses BOOTTIME clock type by default, if it is not mentioned in
the TracePacket. For data sources that use different clock type,
they need to specify it in the TracePacket. SurfaceFlinger uses
MONOTONIC type but this wasn't specified previously, resulting in
misaligned events on the trace.

Test: Take a trace, disconnect the USB cable, turn off the screen,
connect the USB cable after 2 minutes and take a trace again
Bug: 145616132

Change-Id: I1579aba874ecfd6bb8fd1f6d04e6cad61453a62d