[Shadows] Create a new composition layer for shadows [6/n]
When generating the list of composition layers, generate one to render
shadows based on the layer settings.

The function passes in  the display view port to calculate shadow
position. When rendering a screenshot, use the layerstack of the
root layer to find the display and its viewport since the shadows
on the screenshot have to match how it was rendered on display.

Pass in display or target output dataspace so shadows do not have to
do any color conversion.

Bug: 136561771
Test: go/wm-smoke
Test: libcompositionengine_test
Change-Id: I89795707f054b6a08dabc278d80ed393a0da0a7e