Add InputReader Integration tests
The InputReader unit tests rely on internal implementation detail of
InputReader, and do not test the threding functionality. We add a new
test suite that creates and interacts with InputReader solely through
the InputReaderInterface, and also verifies its threading functionality
through the start() and stop() methods.

These integration tests use uinput to emulate input devices, and emit
events, which are verified through the InputListener interface.

Bug: 144595884
Test: atest inputflinger_tests
Change-Id: If51b35cc46180836e60fd28d50c092396c163034