SF: VSyncReactor ensure period change confirmation
(Code being corrected is flagged off, no production problems)

When a period change is happening, system turns on hwvsync and uses
the first vsync signal timed at the new vsync rate as confirmation
that the rate has indeed changed (and then things like modifying the
offsets occurs).

If a presentation fence was added to the VSyncReactor system before
the first hwvsync sample arrived, this would result in the period never
being confirmed, and would result in incorrect offsets and period
estimations, resulting in bad jank scores.

Test: 3 unit tests
Test: uibench run
Fixes: 147321414

Change-Id: Idbc87529e60333df5180de585dbba3525396562b